Split into multiple and normal
