+++ src/js/components/Select/__tests__/SelectMultiple-test.js
@@ -1,7 +1,10 @@
 import React from 'react';
 import 'jest-styled-components';
 import renderer from 'react-test-renderer';
-import { cleanup, render, fireEvent, act } from '@testing-library/react';
+import { cleanup, render, fireEvent } from '@testing-library/react';
+import { axe } from 'jest-axe';
+import 'jest-axe/extend-expect';
+import 'regenerator-runtime/runtime';
 
 import { createPortal, expectPortal } from '../../../utils/portal';
 
@@ -13,87 +16,15 @@ describe('Select Controlled', () => {
   beforeEach(createPortal);
   afterEach(cleanup);
 
-  test('0 value', () => {
-    const component = renderer.create(
-      <Select
-        id="test-select"
-        placeholder="test select"
-        options={[0, 1]}
-        value={0}
-      />,
+  test('should not have accessibility violations', async () => {
+    const { container } = render(
+      <Grommet>
+        <Select options={['one', 'two', 'three']} a11yTitle="test" multiple />
+      </Grommet>,
     );
-    expect(component.toJSON()).toMatchSnapshot();
-  });
-
-  test('search', () => {
-    jest.useFakeTimers();
-    const onSearch = jest.fn();
-    const { getByPlaceholderText, container } = render(
-      <Select
-        id="test-select"
-        placeholder="test select"
-        options={['one', 'two']}
-        onSearch={onSearch}
-        // need value here to show how options/value changes on search
-        value="two"
-      />,
-    );
+    const results = await axe(container);
     expect(container.firstChild).toMatchSnapshot();
-
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    // advance timers so select can open
-    act(() => {
-      jest.advanceTimersByTime(200);
-    });
-    // snapshot on search box
-    expectPortal('test-select__drop').toMatchSnapshot();
-    expect(document.activeElement).toMatchSnapshot();
-    // add content to search box
-    fireEvent.change(document.activeElement, { target: { value: 'o' } });
-    expect(document.activeElement).toMatchSnapshot();
-    expect(onSearch).toBeCalledWith('o');
-  });
-
-  test('select an option with complex options', () => {
-    const onChange = jest.fn();
-    const { getByText, container } = render(
-      <Select
-        id="test-select"
-        plain
-        value={<span>one</span>}
-        options={[{ test: 'one' }, { test: 'two' }]}
-        onChange={onChange}
-      >
-        {option => <span>{option.test}</span>}
-      </Select>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-
-    fireEvent.click(getByText('one'));
-
-    // pressing enter here nothing will happen
-    fireEvent.click(
-      document.getElementById('test-select__drop').querySelector('button'),
-    );
-    expect(onChange).toBeCalledWith(
-      expect.objectContaining({ value: { test: 'one' } }),
-    );
-    expect(window.scrollTo).toBeCalled();
-  });
-
-  test('size', () => {
-    const component = renderer.create(
-      <Select
-        id="test-select"
-        size="large"
-        options={['one', 'two']}
-        selected={[]}
-        value={[]}
-        onChange={() => {}}
-      />,
-    );
-    expect(component.toJSON()).toMatchSnapshot();
+    expect(results).toHaveNoViolations();
   });
 
   test('multiple', () => {
@@ -175,69 +106,6 @@ describe('Select Controlled', () => {
     );
     expect(onChange).toBeCalledWith(expect.objectContaining({ value: [] }));
   });
-  ['small', 'medium', 'large'].forEach(dropHeight => {
-    test(`${dropHeight} drop container height`, () => {
-      const { getByPlaceholderText } = render(
-        <Select
-          id="test-select"
-          size="large"
-          options={['one', 'two']}
-          selected={[]}
-          value={[]}
-          onChange={() => {}}
-          dropHeight={dropHeight}
-          placeholder="test select"
-        />,
-      );
-      fireEvent.click(getByPlaceholderText('test select'));
-      expect(document.activeElement).toMatchSnapshot();
-    });
-  });
-
-  test('onChange without valueKey', () => {
-    const onChange = jest.fn();
-    const Test = () => {
-      const [value] = React.useState();
-      return (
-        <Select
-          id="test-select"
-          placeholder="test select"
-          labelKey="name"
-          value={value}
-          options={[
-            {
-              id: 1,
-              name: 'Value1',
-            },
-            {
-              id: 2,
-              name: 'Value2',
-            },
-          ]}
-          onChange={onChange}
-        />
-      );
-    };
-    const { getByPlaceholderText, getByText, container } = render(
-      <Grommet>
-        <Test />
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    expectPortal('test-select__drop').toMatchSnapshot();
-
-    fireEvent.click(getByText('Value1'));
-    expect(onChange).toBeCalledWith(
-      expect.objectContaining({
-        value: {
-          id: 1,
-          name: 'Value1',
-        },
-      }),
-    );
-  });
 
   test('multiple onChange without valueKey', () => {
     const onChange = jest.fn();
@@ -302,52 +170,6 @@ describe('Select Controlled', () => {
             name: 'Value2',
           },
         ],
-      }),
-    );
-  });
-
-  test('onChange with valueKey string', () => {
-    const onChange = jest.fn();
-    const Test = () => {
-      const [value] = React.useState();
-      return (
-        <Select
-          id="test-select"
-          placeholder="test select"
-          labelKey="name"
-          valueKey="id"
-          value={value}
-          options={[
-            {
-              id: 1,
-              name: 'Value1',
-            },
-            {
-              id: 2,
-              name: 'Value2',
-            },
-          ]}
-          onChange={onChange}
-        />
-      );
-    };
-    const { getByPlaceholderText, getByText, container } = render(
-      <Grommet>
-        <Test />
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    expectPortal('test-select__drop').toMatchSnapshot();
-
-    fireEvent.click(getByText('Value1'));
-    expect(onChange).toBeCalledWith(
-      expect.objectContaining({
-        value: {
-          id: 1,
-          name: 'Value1',
-        },
       }),
     );
   });
@@ -495,44 +317,6 @@ describe('Select Controlled', () => {
     );
     expect(container.firstChild).toMatchSnapshot();
     fireEvent.click(getByPlaceholderText('test select'));
-    expectPortal('test-select__drop').toMatchSnapshot();
-  });
-
-  test('disabled key', () => {
-    jest.useFakeTimers();
-    const Test = () => {
-      const [value] = React.useState();
-      return (
-        <Select
-          id="test-select"
-          placeholder="test select"
-          labelKey="name"
-          valueKey="id"
-          value={value}
-          disabledKey="disabled"
-          options={[
-            {
-              id: 1,
-              name: 'Value1',
-              disabled: true,
-            },
-            {
-              id: 2,
-              name: 'Value2',
-              disabled: false,
-            },
-          ]}
-        />
-      );
-    };
-    const { getByPlaceholderText, container } = render(
-      <Grommet>
-        <Test />
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-    fireEvent.click(getByPlaceholderText('test select'));
-    jest.advanceTimersByTime(200);
     expectPortal('test-select__drop').toMatchSnapshot();
   });
   window.scrollTo.mockRestore();
